refactor: move objectives files into package objective
