+++ src/main/scala/model/map/RoutePointsManager.scala
@@ -1,15 +1,17 @@
 package model.map
 
+import model.objective.Points
+
 /** Trait that represents a manager of route points. It provides a way to calculate the points given by each route. */
 trait RoutePointsManager:
-  import RoutePointsManager.Points
-
-  /** Returns the number of points given by the specified route. */
+  /** Returns the number of points given by the specified route.
+    *
+    * @return
+    *   the number of points given by the specified route
+    */
   extension (route: Route) def points: Points
 
 object RoutePointsManager:
-  export model.player.Points
-
   /** Creates a [[RoutePointsManager]] that calculates the points based on the length of the route.
     *
     * @return
