+++ src/main/scala/model/objective/Objective.scala
@@ -1,4 +1,4 @@
-package model.player
+package model.objective
 
 export model.map.GameMap.CityName as City
