+++ src/test/scala/model/objective/ObjectiveTest.scala
@@ -1,4 +1,4 @@
-package model.player
+package model.objective
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
