+++ src/test/scala/model/player/PlayerTest.scala
@@ -1,5 +1,6 @@
 package model.player
 
+import model.objective.{ObjectiveCompletion, ObjectiveWithCompletion, Points}
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
